done refactoring, lots of modularity (see experiments)
