+++ .gitignore
@@ -5,3 +5,8 @@ plots/*
 **__pycache__**
 data/*
 *.code-workspace
+**/outputs/**
+**outputs**
+**/data/**
+**/archive/**
+**/*egg*/**
