+++ minma/__init__.py
@@ -1,0 +1,20 @@
+"""
+TODO: assign cleaner package exports for public functions only
+"""
+
+from . import data
+from . import dsp 
+from . import run_experiments
+from . import glm 
+from . import viz
+from . import models
+
+
+__all__ = [
+    'data',
+    'dsp',
+    'main',
+    'glm',
+    'viz',
+    'models'
+]
