+++ minma/notes/meet_John.md
