+++ minma/notes/model_notes.md
@@ -1,0 +1,9 @@
+
+Small Modeling Ideas
+- XGBOOST
+- Random Forest
+- Combining Features
+- Combining GLM and other model with ensemble / bagging
+- Bayesian Ridge Regression (classificaiton?)
+- Gaussian Process
+-
