+++ setup.py
@@ -1,0 +1,21 @@
+#!/usr/bin/env python3
+
+import io
+import os
+import re
+from setuptools import find_packages, setup
+
+
+# Run the setup
+setup(
+    name="minma",
+    description="",
+    version="0.0.1",
+    author="",
+    url="",
+    author_email="",
+    license="MIT",
+    classifiers=["Programming Language :: Python :: 3"],
+    python_requires=">=3.6",
+    install_requires=[],
+)
